start by adding the stylesheet converter
This turns out much simpler as in original RenderLegend as I plan to
read data from a dummy PostGIS database instead of OSM XML directly.

This is for two reasons:

* Reading from the same datasource type makes rewriting the SQL
  statements in the layers unnecessary. Especially given that the
  original RenderLegend SQL "parser" is not able to parse some
  of the SQL statements in the current OpenStreetMap Carto style
  anyway, like e.g those using UNION, and as fully parsing SQL
  is hard, it's probably better to get rid of this approach
  completely.

* The OSM input plugin is no longer part of core mapnik, and is
  usually missing from distribution packages. While the OGR
  input plugin can also read OSM XML, it behaves subtly different
  enough to make porting the original code hard.

So we just replace the "dbname" to read from, and shift the conversion
problem from "rewrite the layer SQL" to "move our generated OSM XML into
a small PostGIS database", for which perfectly working tools already
exist, or for which we could even generate SQL WKT gis data directly,
still at a much lower complexity level than trying to build a general
SQL parser.

One item not covered yet: references to image files in the style file
are usually relative to the files path. So for now we have to have
original and rewritten style files in the same directory. Ideally
such relative paths should be rewritten to become absolute paths
referring to the original files, so that we can put the rewritten
style whereever we want.